tehtävä 2

+++ tehtävät 2/tehtävä 4.py
@@ -1,4 +1,4 @@
-width = int(input("Mikä on suorakolmion kannan pituus?"))
-height = int(input("Mikä on suorakolmion korkeus?"))
+width = float(input("Mikä on suorakolmion kannan pituus?"))
+height = float(input("Mikä on suorakolmion korkeus?"))
 print("Suorakulmion piiri on", width*2+height*2)
 print("suorakulmion pinta-ala", width*height)
